% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtpriors.R, R/gtpriors_mpfr.R
\name{pkn_ngg}
\alias{pkn_ngg}
\alias{pkn_ngg.mpfr}
\title{Computes the Gibbs type prior number of cluster}
\usage{
pkn_ngg(k, n, beta, sigma, prec = 53, rec = F)

pkn_ngg.mpfr(k, n, beta, sigma, prec = 53)
}
\arguments{
\item{k}{The value for k}

\item{n}{The value for n}

\item{beta}{The for beta}

\item{sigma}{The value for sigma}

\item{prec}{The number of bits of precision used during internal computation}

\item{rec}{Recursive formulation of Cnk}
}
\value{
The density in k for the prior distribution on the number of clusters for a Normalised Generalised Gamma process
}
\description{
Computes the Gibbs type prior number of cluster
}
\examples{
pkn_ngg(1:100, 100, 1.2, 0.8, prec=1000)

}
