% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtpriors.R, R/gtpriors_mpfr.R
\name{cnk_ngg}
\alias{cnk_ngg}
\alias{cnk_ngg.mpfr}
\title{Computes the generalized binomial coefficients for Gibbs type prior number of cluster}
\usage{
cnk_ngg(n, k, sigma, prec = 53, rec = F)

cnk_ngg.mpfr(n, k, sigma, prec = 53)
}
\arguments{
\item{n}{The value for n}

\item{k}{The value for k}

\item{sigma}{The value for sigma}

\item{prec}{The number of bits of precision used during internal computation}

\item{rec}{Use the recursive formula}
}
\value{
The value of the generalized binomial coefficient
}
\description{
Computes the generalized binomial coefficients for Gibbs type prior number of cluster
}
