% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGibbsTypePriors.R
\name{pkn_ngg}
\alias{pkn_ngg}
\title{Computes the prior number of clusters for a Normalised Generalised Gamma process}
\usage{
pkn_ngg(k, n, beta, sigma, prec = 53)
}
\arguments{
\item{k}{The value for k}

\item{n}{The value for n}

\item{beta}{The value for beta}

\item{sigma}{The value for sigma}

\item{prec}{The number of bits of precision used during internal computation}
}
\value{
The density in k for the prior distribution on the number of clusters for a Normalised Generalised Gamma process
}
\description{
Computes the prior number of clusters for a Normalised Generalised Gamma process
}
