% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGibbsTypePriors.R
\name{vnk_ngg}
\alias{vnk_ngg}
\title{Computes the weights for a Normalised Generalised Gamma process}
\usage{
vnk_ngg(n, k, beta, sigma, prec = 53)
}
\arguments{
\item{n}{The value for n}

\item{k}{The value for k}

\item{beta}{The value for beta}

\item{sigma}{The value for sigma}

\item{prec}{The number of bits of precision used during internal computation}
}
\value{
The Vnk weights for a Normalised Generalised Gamma process
}
\description{
Computes the weights for a Normalised Generalised Gamma process
}
